tests: Reorganize pytest utils and fixtures.
- Move assert middleware and helper function to utils.
- Move client fixture to conftest.py.
- Edit test_auth.py in order to go with that changes.
